+++ shop/src/main/java/cn/qingchengfit/shop/ui/items/inventory/IInventoryRecordData.java
@@ -18,5 +18,5 @@ public interface IInventoryRecordData {
 
   int getOffset();
 
-  int getInventorCount();
+  long getInventorCount();
 }
