+++ shop/src/main/java/cn/qingchengfit/shop/ui/items/product/IProductItemData.java
@@ -18,7 +18,7 @@ public interface IProductItemData {
 
   int getProductSales();
 
-  int getProductInventory();
+  long getProductInventory();
 
   int getProductPriority();
